AboutUs, Attachment and AboutUsAttachment features Developed

+++ HealthTourist.Application/Features/AboutUsPage/Commands/DeleteAboutUs/DeleteAboutUsCommandHandler.cs
@@ -1,4 +1,6 @@
 using HealthTourist.Application.Contracts.AboutUsPage;
+using HealthTourist.Common.Constants.AboutUs;
+using HealthTourist.Common.Exceptions;
 using MediatR;
 
 namespace HealthTourist.Application.Features.AboutUsPage.Commands.DeleteAboutUs;
@@ -11,6 +13,10 @@ public class DeleteAboutUsCommandHandler(IAboutUsRepository aboutUsRepository)
     public async Task<Unit> Handle(DeleteAboutUsCommand request, CancellationToken cancellationToken)
     {
         // Validate incoming request
+        var validator = new DeleteAboutUsCommandValidator(aboutUsRepository);
+        var validationResult = await validator.ValidateAsync(request, cancellationToken);
+        if (validationResult.Errors.Count != 0)
+            throw new BadRequestException(AboutUsExceptionConstants.BadRequestExceptionMessage, validationResult);
         
         // Delete
         await aboutUsRepository.DeleteAsync(request.Id);
